payment_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mercadopago==2.2.0
pydantic==2.5.0
python-dotenv==1.0.0
pika==1.3.2
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7